+++ utils/convert.go
@@ -13,27 +13,11 @@ import (
 	"unsafe"
 )
 
-const MaxStringLen = 0x7fff0000 // Maximum string length for UnsafeBytes. (decimal: 2147418112)
-
 // UnsafeString returns a string pointer without allocation
 //
 //nolint:gosec // unsafe is used for better performance here
 func UnsafeString(b []byte) string {
 	return *(*string)(unsafe.Pointer(&b))
-}
-
-// UnsafeBytes returns a byte pointer without allocation.
-// String length shouldn't be more than 2147418112.
-//
-//nolint:gosec // unsafe is used for better performance here
-func UnsafeBytes(s string) []byte {
-	if s == "" {
-		return nil
-	}
-
-	return (*[MaxStringLen]byte)(unsafe.Pointer(
-		(*reflect.StringHeader)(unsafe.Pointer(&s)).Data),
-	)[:len(s):len(s)]
 }
 
 // CopyString copies a string to make it immutable
